プレイヤーのHPバーを実装した

+++ Assets/ScriptableObjects/DamageObjectParameter.cs
@@ -6,8 +6,8 @@ namespace Assets.ScriptableObjects
     [CreateAssetMenu(menuName = "ScriptableObjects/DamageObjectParameter")]
     public class DamageObjectParameter : ScriptableObject
     {
-        [SerializeField] private float _attackPower;
-        public float AttackPower => _attackPower;
+        [SerializeField] private int _attackPower;
+        public int AttackPower => _attackPower;
         [SerializeField] private float _readyTime;
         public float ReadyTime => _readyTime;
         [SerializeField] private float _damagingTime;
