+++ Assets/Scripts/Objects/IObject.cs
@@ -8,7 +8,7 @@ namespace Assets.Scripts.Objects
         public void SetReady();
         public bool IsDamaging();
         public void DamageTo(IObject obj);
-        public void TakeDamage(float damage);
+        public void TakeDamage(int damage);
         public (Vector3 minImPos3, Vector3 maxImPos3) GetImPos3s(); 
         public Vector3 GetRePos3();
         public void DestroyDeadObject();
